% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsmlr.R
\docType{data}
\name{bin_data}
\alias{bin_data}
\title{Ordinal value and default indicator data}
\format{
A data frame with 200,000 rows and 2 variables:
\describe{
  \item{score}{the score is the ordinal value that indicates the relative
default risk of the obligor. The higher the score, the lower is the predicted
default risk as proposed by the rank ordering model}
  \item{default}{a default indicator where 1 implies default status (in the
next 12-month period) and 0 indicates non-default status}
}
}
\usage{
bin_data
}
\description{
A dataset containing the scores and default indicator of 200,000 obligors.
}
\keyword{datasets}
