% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_binning.R
\name{create_initial_bins}
\alias{create_initial_bins}
\title{Create initial bins}
\usage{
create_initial_bins(
  data = NULL,
  threshold = NULL,
  ordinal_value = NULL,
  default_ind = NULL
)
}
\arguments{
\item{data}{The input data that contains \code{ordinal_value} and
\code{default_ind}.}

\item{threshold}{An integer that defines the minimum number of defaults in
each bin. Typically needs a reasonable number of defaults to ensure the
homogeneity test can be executed with some power.}

\item{ordinal_value}{A numeric ordinal value that rank order each obligor's
relative default risk. For example, a numeric score or numeric rating
information.}

\item{default_ind}{An integer, or double, representing the default status.
The state of default must be indicated by 1 and non-default by 0.}
}
\value{
A list of bins. Each list component in the returned list is a bin (of
  class bin). The list is returned in ascending order by min_score (see link
  to constructor below), and each bin contains at least \code{threshold}
  number of defaults.
}
\description{
Creates a set of initial bins.
}
\details{
The current (only) method is based on an 'equal height' approach where a new
bin is created as soon as the number of defaults is equal to (or larger) than
a user defined \code{threshold} level. Thus, the term 'equal height' refers
to roughly equal number of defaults in each bin. The data is processed based
on the user supplied variables holding the \code{ordinal_value} and the
\code{default_ind}. The state of default in \code{default_ind} is generally 1
if the obligor defaults in the next 12-month period. To avoid bias in the
default rate calculation, only current non-default obligors should be
included in \code{data}. This type of data pre-processing is left to the user
before calling \code{\link{create_initial_bins}}. Similary, the rank-order
information in \code{ordinal_value} must be invariant to sorting. Thus, it
may be required to convert characters/factors of ratings to numerical values
before calling \code{\link{create_initial_bins}}.
}
\examples{
# create a set of initial bins
bins <- create_initial_bins(bin_data, 30, "score", "default")
bins_df <- dplyr::bind_rows(bins)
plot(x = bins_df$mid_score, y = log(bins_df$odds), type = "p",
     col = "lightblue", cex = 1.5, pch = 20,
     ylab = "log(odds)", xlab = "score")
}
\seealso{
The (helper) constructor to bin objects, \code{\link{bin}}. See also
  \code{\link{reduce_bins}}.
}
