% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_monotonic}
\alias{is_monotonic}
\title{Verify monotonicity condition of default rates}
\usage{
is_monotonic(list_of_bins, decreasing = TRUE, verbose = FALSE)
}
\arguments{
\item{list_of_bins}{A list of bins. Every bin object contains a default rate.}

\item{decreasing}{A boolean. Should the monotonicity test verify a decrease
or increase in default rates?}

\item{verbose}{A boolean. If set to TRUE, prints a matrix of default rates
along with a binary indicator displaying 1 where the monotonic
relationships fails to hold.}
}
\value{
A boolean. TRUE indicates a monotonic development in default rates
  across all bins in \code{list_of_bins}.
}
\description{
Verify if a list of bins is monotonic in terms of default rates. The function
\code{is_monotonic} returns a boolean indicating TRUE if the default rates
exhibit a monotonic trend.
}
\details{
A list of bins of length N contains N default rates. The monotonic condition
tested by this function examines whether the N default rates decrease (or
increase) for each consecutive default rate in the list. The test is
non-strict, meaning that if the default rates between any two adjacent bins
are equal this is not considered a breach of monotonicity.
}
\examples{
# create a set of initial bins and verify where monotonicity fails
bins <- create_initial_bins(bin_data, 30, "score", "default")
is_monotonic(bins, verbose = TRUE)
}
