% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{merge_list_of_bins}
\alias{merge_list_of_bins}
\title{Merge list of bins}
\usage{
merge_list_of_bins(list_of_bins, idx)
}
\arguments{
\item{list_of_bins}{A list of bins.}

\item{idx}{An integer (index) vector indicating which bins in
\code{list_of_bins} to merge. The index vector must be sorted in ascending
order without any gaps.}
}
\value{
A list of bins. Each list component in the returned list is a bin (of
  class bin). Total length equal to \code{length(list_of_bins) - length(idx)
  + 1}.
}
\description{
A merge function to collapse bin objects. This function is used by other
functions in the binsmlr package, but it is also useful on a standalone basis
when manual binning is required (enforce monotonicity across bins in terms of
default rates).
}
\examples{
# example of manual binning
autobins <- autobin(bin_data, 30, "score", "default", 12, 0.01)
length(autobins)
is_monotonic(autobins, verbose = TRUE)
man_bins <- merge_list_of_bins(autobins, 3:4)
length(man_bins)
is_monotonic(man_bins)
}
